+++ taskiq/abc/__init__.py
@@ -1,4 +1,5 @@
 """Abstract classes for taskiq."""
+
 from taskiq.abc.broker import AsyncBroker
 from taskiq.abc.result_backend import AsyncResultBackend
