+++ taskiq/middlewares/__init__.py
@@ -1,6 +1,5 @@
 """Taskiq middlewares."""
 
-
 from .prometheus_middleware import PrometheusMiddleware
 from .simple_retry_middleware import SimpleRetryMiddleware
 from .smart_retry_middleware import SmartRetryMiddleware
