+++ taskiq/schedule_sources/__init__.py
@@ -1,4 +1,5 @@
 """Package for schedule sources."""
+
 from taskiq.schedule_sources.label_based import LabelScheduleSource
 
 __all__ = [
