+++ taskiq/serializers/__init__.py
@@ -1,4 +1,5 @@
 """Taskiq serializers."""
+
 from .cbor_serializer import CBORSerializer
 from .json_serializer import JSONSerializer
 from .msgpack_serializer import MSGPackSerializer
